# Some micro-perf-optimizations