--- a/csFastFloat/AssemblyInfo.cs
+++ b/csFastFloat/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("TestcsFastFloat")]
+[assembly: InternalsVisibleTo("Benchmark")]
+
+#if NET5_0
+[module: SkipLocalsInit]
+#endif