--- a/csFastFloat/FastDoubleParser.cs
+++ b/csFastFloat/FastDoubleParser.cs
@@ -239,15 +239,7 @@ namespace csFastFloat
     }
 
 
-
-
-    static readonly byte[] powers = {
-        0,  3,  6,  9,  13, 16, 19, 23, 26, 29, //
-        33, 36, 39, 43, 46, 49, 53, 56, 59,     //
-    };
-
-
-        internal static AdjustedMantissa ComputeFloat(DecimalInfo d)
+    internal static AdjustedMantissa ComputeFloat(DecimalInfo d)
     {
       AdjustedMantissa answer = new AdjustedMantissa();
       if (d.num_digits == 0)
@@ -289,7 +281,7 @@ namespace csFastFloat
       while (d.decimal_point > 0)
       {
         uint n = (uint)(d.decimal_point);
-        int shift = (n < num_powers) ? powers[n] : max_shift;
+        int shift = (n < num_powers) ? Constants.get_powers(n) : max_shift;
 
         d.decimal_right_shift(shift);
         if (d.decimal_point < -Constants.decimal_point_range)
@@ -318,7 +310,7 @@ namespace csFastFloat
         else
         {
           uint n = (uint)(-d.decimal_point);
-          shift = (n < num_powers) ? powers[n] : max_shift;
+          shift = (n < num_powers) ? Constants.get_powers(n) : max_shift;
         }
 
         d.decimal_left_shift(shift);