--- a/csFastFloat/FastDoubleParser.cs
+++ b/csFastFloat/FastDoubleParser.cs
@@ -36,14 +36,11 @@ namespace csFastFloat
 
     public static unsafe double ToFloat(bool negative, AdjustedMantissa am)
     {
-      double d;
       ulong word = am.mantissa;
       word |= (ulong)(am.power2) << DoubleBinaryConstants.mantissa_explicit_bits;
       word = negative ? word | ((ulong)(1) << DoubleBinaryConstants.sign_index) : word;
 
-      Unsafe.CopyBlockUnaligned(&d, &word, sizeof(double));
-
-      return d;
+      return BitConverter.Int64BitsToDouble((long)word);
     }    
     
     public  static double FastPath(ParsedNumberString pns)