--- a/csFastFloat/csFastFloat.csproj
+++ b/csFastFloat/csFastFloat.csproj
@@ -13,6 +13,7 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Title>csFastFloat : a fast and accurate float parser</Title>
     <PackageLicenseFile>README.md</PackageLicenseFile>
+    <LangVersion>9</LangVersion>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">