Skimmed through the code base (really impressive algorithm) and fixed some low hanging fruits which jumped into my eye.
Basically it's exercising some C# and JIT-tricks, as well some new APIs in .NET 5, especially for array-accesses where the bound-checks are avoided. They can sum-up if done for each access to an array or span.

Please see individual commits, they have links in the message that help to explain / demonstrate the difference in generated asm-code.

### Benchmarks

#### This PR + #54

Based on caaccaa9bd9ea88a7e1491bde59de653bec8e6c0 with #54 merged into that commit.

```
|                   Method |        FileName |      Mean |     Error |    StdDev |       Min | Ratio | MFloat/s |     MB/s |
|------------------------- |---------------- |----------:|----------:|----------:|----------:|------:|---------:|---------:|
|  FastFloat.ParseDouble() | data/canada.txt |  6.263 ms | 0.0587 ms | 0.0549 ms |  6.166 ms |  0.16 |    18,02 |   356,33 |
| ParseNumberString() only | data/canada.txt |  3.487 ms | 0.0312 ms | 0.0277 ms |  3.454 ms |  0.09 |    32,18 |   636,14 |
|           Double.Parse() | data/canada.txt | 39.207 ms | 0.7458 ms | 0.6612 ms | 38.294 ms |  1.00 |     2,90 |    57,37 |
|                          |                 |           |           |           |           |       |          |          |
|  FastFloat.ParseDouble() |   data/mesh.txt |  2.329 ms | 0.0178 ms | 0.0148 ms |  2.301 ms |  0.34 |    31,73 |   300,27 |
| ParseNumberString() only |   data/mesh.txt |  1.593 ms | 0.0173 ms | 0.0145 ms |  1.572 ms |  0.23 |    46,44 |   439,43 |
|           Double.Parse() |   data/mesh.txt |  6.792 ms | 0.1302 ms | 0.1599 ms |  6.615 ms |  1.00 |    11,04 |   104,46 |

```

#### Current master

Current master is 788081cb2bb7a8846096357a15b5a7fe86773a70

```
|                   Method |        FileName |      Mean |     Error |    StdDev |       Min | Ratio | MFloat/s |     MB/s |
|------------------------- |---------------- |----------:|----------:|----------:|----------:|------:|---------:|---------:|
|  FastFloat.ParseDouble() | data/canada.txt |  7.085 ms | 0.1272 ms | 0.0993 ms |  6.849 ms |  0.18 |    16,23 |   320,80 |
| ParseNumberString() only | data/canada.txt |  3.783 ms | 0.0733 ms | 0.0784 ms |  3.672 ms |  0.09 |    30,26 |   598,31 |
|           Double.Parse() | data/canada.txt | 39.955 ms | 0.6315 ms | 0.5598 ms | 39.123 ms |  1.00 |     2,84 |    56,16 |
|                          |                 |           |           |           |           |       |          |          |
|  FastFloat.ParseDouble() |   data/mesh.txt |  2.487 ms | 0.0323 ms | 0.0287 ms |  2.449 ms |  0.37 |    29,81 |   282,12 |
| ParseNumberString() only |   data/mesh.txt |  1.700 ms | 0.0324 ms | 0.0287 ms |  1.660 ms |  0.25 |    43,98 |   416,15 |
|           Double.Parse() |   data/mesh.txt |  6.678 ms | 0.1084 ms | 0.0961 ms |  6.541 ms |  1.00 |    11,16 |   105,65 |
```