@lemire 2021-02-24
I expect that the sizeable performance gains can be attributed to two reasons: `Unsafe.CopyBlockUnaligned` to avoid input validation and the non-bound-checking  `exact_power_of_ten`. But I am surprised that it can boost the canada performance by ~10% so there might be other important effects that I am missing. (I am aware that there are other optimizations in this PR, but I expect that they account for a small percentage of the gain.)