@gfoidl 2021-02-24
I didn't measure each step, but from gut feeling I'd say

1. removing of the bound-checks (it's the `cmp` and `jmp` combo) which also makes the code (especially loops) tighter*
2. `is_integer` has a `movzx` less, but it's called (note: it's inlined) very often, so it will sum-up
3. `Unsafe.CopyBlockUnaligned`**

\* the branch predictor will do a good job -- at least I assume so -- but tighter loops may be a reason for the boost, maybe we are lucky and the loops will be properly aligned. I don't know the current loop-alignment rules in the JIT, and there were some recent changes, so I'm not up to date here.

\** maybe I'm wrong here and the effect is greater, as you assume. [It compiles](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgCUBXAOwwEt8YLAMIR8AB14AbGFADKMgG68wMXAG4aNYgGYG3XNgBmMJqQZCGAbxoNbTXcRQMAsgApHAKga4IHKCrQmFC8AE1UMAEoGGztrajsEhgBVLgNjYQgxAE8AIUkIMABrFOxJXgBzLhgQ1zDcDECfPwDvXgAvGAhDWt9gaQiIjXi7AF8Y23H7IJdyd2DvX38YQM8GOsjJuMS7HI5DYygWZxh8aCyRbNcmpcD1xvbO7pDe/vuOrp6OPpgBoYSx6gjIA=) down to mem -> reg -> mem, so ideal.