@gfoidl 2021-02-24
Ah, almost forgot. Removing the bound-checks has another effect.
In .NET the length of the array is stored before the first element, and for a bound-check the length is required. So the length and the accessed element may reside on different cachelines. 

Would need to investigate how much this effect is actually.
For parallel code this effect (resuling in false sharing) can be seen clearly (https://gist.github.com/gfoidl/346ea7472f06e2024b8bdc9c598e547e#gistcomment-2773628).