@gfoidl 2021-02-25
Next perf steps would be looking into it with VTune, PerfView, etc.
With looking at the compiled assembly code I can't spot anything (at the moment), which can be optimized w/o changing the JIT. Maybe a few littel steps may still arise.