@gfoidl 2021-02-25
## VTune

![grafik](https://user-images.githubusercontent.com/5755834/109141407-ff128600-775d-11eb-9a80-ad57637d34d0.png)

### ParsedNumberString.ParseNumberString

If I'm not off it's 

![grafik](https://user-images.githubusercontent.com/5755834/109144168-2028a600-7761-11eb-8d01-960ffe0c4260.png)

### FastDoubleParser.ComputeFloat

![grafik](https://user-images.githubusercontent.com/5755834/109142757-8b717880-775f-11eb-827d-1cb7af1f7140.png)

### FastDoubleParser.ParseNumber

![grafik](https://user-images.githubusercontent.com/5755834/109142573-52390880-775f-11eb-8d95-13d7074a5a98.png)