@lemire 2021-02-25
@gfoidl You would think that the `is_integer` optimization would count for a lot, but if you look at ParseNumberString results, at least according to your numbers, they are not helped that much. So the performance boost remains a bit mysterious to me.

Well. We are talking about ~5% to ~15% so it is not like these sort of optimizations have fully predictable results from first principles.

(Note : ~5% to ~15% is huge for the use case here. I am not dismissing the achievement.)