@lemire 2021-02-25
@CarlVerret I recommend merging. The `is_integer` code is a tad ugly (😄), but it appears to be justified.