@gfoidl 2021-02-25
I'm just running benchmarks, looks promising.

The `is_integer` won't change the needle much, it safes "just" one redundant add with a constant. The cpu's execution will do this in "parallel" anyway, but I didn't like this redundancy and it's a bit less instructions to fetch & decode.

> So the performance boost remains a bit mysterious to me.

😉 On the other side for me it's quite obvious, but I'm working with C# / .NET for quite a long time, so some of the inner workings are familiar to me.

Overall the biggest player for the perf-improvements is the removal of bound-checks.