@gfoidl 2021-02-25
Latest numbers:

```
|                   Method |        FileName |      Mean |     Error |    StdDev |       Min | Ratio | MFloat/s |     MB/s |
|------------------------- |---------------- |----------:|----------:|----------:|----------:|------:|---------:|---------:|
|  FastFloat.ParseDouble() | data/canada.txt |  6.085 ms | 0.1034 ms | 0.0864 ms |  5.985 ms |  0.15 |    18,57 |   367,08 |
| ParseNumberString() only | data/canada.txt |  3.368 ms | 0.0463 ms | 0.0411 ms |  3.319 ms |  0.08 |    33,48 |   661,89 |
|           Double.Parse() | data/canada.txt | 39.539 ms | 0.7062 ms | 1.1404 ms | 38.550 ms |  1.00 |     2,88 |    56,99 |
|                          |                 |           |           |           |           |       |          |          |
|  FastFloat.ParseDouble() |   data/mesh.txt |  2.274 ms | 0.0221 ms | 0.0196 ms |  2.236 ms |  0.35 |    32,66 |   309,08 |
| ParseNumberString() only |   data/mesh.txt |  1.571 ms | 0.0247 ms | 0.0231 ms |  1.544 ms |  0.24 |    47,30 |   447,64 |
|           Double.Parse() |   data/mesh.txt |  6.440 ms | 0.0550 ms | 0.0459 ms |  6.379 ms |  1.00 |    11,45 |   108,33 |
```

It's 0f59fe8261cfa273bf71363525da762e8782f94e with #54 merged into it.