@lemire 2021-02-25
> The cpu's execution will do this in "parallel" anyway

You might be pessimistic in the value of your optimization. I do not not have the numbers for C#, but in C++, this code is at about 280 instructions per float... so saving, say, 20 instructions, would be 5% to 10% of the total instruction count. Of course, instruction counts do not translate into better performance... normally... but if you look at the number of instructions retired per cycle for this code, in C++, we pretty much max out the width of the processor (>=4 ins. per cycle on Intel/AMD, 8 ins. per cycle on ARM M1). Look at the `ParseNumberString` numbers. These do not benefit for the bound checking optimization when looking up precomputed floats (exact_power_of_ten). For canada, you are at min of 3.319 ms whereas the main branch is at  3.672 ms. So you basically have a 10% boost. For mesh, your min is at 1.544 ms starting from  1.660 ms, so ~7%. And mesh has shorter strings.

If we were assumed that it was ASCII instead of the pesky UTF-16, we could go faster even !