@lemire 2021-02-25
> Overall the biggest player for the perf-improvements is the removal of bound-checks.


Let us take canada. You went from 6.849 ms to 5.985 ms a saving of 0.86 ms. The bound checking part should not benefit ParseNumberString and it went from 3.319 ms to 3.672 ms, a saving of 0.35 ms. So I would expect that 0.5 ms of the gain (or ~2/3) is due to the unsafe copy  and the non-bound-checking exact_power_of_ten. Now, mesh is going to benefit much more from the non-bound-checking exact_power_of_ten because it relies entirely on the fast path. Yet of the two sets, you see the larger benefits with canada (14%) as opposed to mesh (9%).


In any case, this is fantastic work. People often say that bound checking is free, and you are definitively showing that it is not case.